IP: some python scripting attempt I had locally
I may work on this some day as I still have that nasty problem when it
comes to exporting my time from Toggl and importing it to Jira.

PHP version works, but it's not error prone at all and uses a mix of PHP
and bash (I really hate it).

:v: